requirements/backend.txt
# ── Core ML ───────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# ── FastAPI & ASGI ────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
python-multipart>=0.0.9

# ── Pydantic v2 ───────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0

# ── Database ──────────────────────────────────────────────────
pymongo>=4.6.0
motor>=3.3.0               # Async MongoDB driver

# ── Cache ─────────────────────────────────────────────────────
redis>=5.0.0

# ── Logging ───────────────────────────────────────────────────
structlog>=24.1.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0              # For async HTTP calls if needed
